[FBNB-99] refactor: Event 구조 변경
