+++ src/main/java/com/prgrms/airbnb/domain/room/event/ChangeReviewInfoListener.java
@@ -1,6 +1,7 @@
-package com.prgrms.airbnb.domain.review.service;
+package com.prgrms.airbnb.domain.room.event;
 
 import com.prgrms.airbnb.domain.common.exception.NotFoundException;
+import com.prgrms.airbnb.domain.review.event.ChangeReviewInfoEvent;
 import com.prgrms.airbnb.domain.room.entity.Room;
 import com.prgrms.airbnb.domain.room.repository.RoomRepository;
 import lombok.extern.slf4j.Slf4j;
@@ -10,20 +11,20 @@ import org.springframework.stereotype.Service;
 
 @Slf4j
 @Service
-public class UpdateReviewInfoListener {
+public class ChangeReviewInfoListener {
 
   private final RoomRepository roomRepository;
 
-  public UpdateReviewInfoListener(RoomRepository roomRepository) {
+  public ChangeReviewInfoListener(RoomRepository roomRepository) {
     this.roomRepository = roomRepository;
   }
 
   @Async
-  @EventListener(UpdateReviewInfoEvent.class)
-  public void handleContextStart(UpdateReviewInfoEvent event) {
+  @EventListener(ChangeReviewInfoEvent.class)
+  public void handleContextStart(ChangeReviewInfoEvent event) {
     Room room = roomRepository.findById(event.getRoomId()).orElseThrow(() -> {
       throw new NotFoundException(this.getClass().getName());
     });
-    room.getReviewInfo().updateReviewInfo(event.getRating());
+    room.getReviewInfo().changeReviewInfo(event.getOldRating(), event.getNewRating());
   }
 }
