+++ src/test/java/com/prgrms/airbnb/domain/review/service/GuestReviewServiceTest.java
@@ -19,6 +19,8 @@ import com.prgrms.airbnb.domain.review.dto.ReviewResponse;
 import com.prgrms.airbnb.domain.review.dto.UpdateReviewRequest;
 import com.prgrms.airbnb.domain.review.entity.Review;
 import com.prgrms.airbnb.domain.review.entity.ReviewImage;
+import com.prgrms.airbnb.domain.review.event.ChangeReviewInfoEvent;
+import com.prgrms.airbnb.domain.review.event.UpdateReviewInfoEvent;
 import com.prgrms.airbnb.domain.review.repository.ReviewRepository;
 import com.prgrms.airbnb.domain.room.entity.Room;
 import com.prgrms.airbnb.domain.room.entity.Room.ReviewInfo;
